Took out AccessoryRarity out of AccessorySO into a new script
