+++ DAT257AgileProject/Assets/Scripts/Accessory/AccessoryRarity.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1acd42d93e86a134489a76c25b0d9301
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
